modify zcc test for tf2
